refactor: edit the code of Invoice.

+++ AquaEngine.API/Sales/Domain/Model/ValueObjects/InvoiceId.cs
@@ -1,3 +1,3 @@
 ﻿namespace AquaEngine.API.Sales.Domain.Model.ValueObjects;
 
-public record InvoiceId(int Id);
+public record InvoiceId(long Id);
